dummy strategy helper moved to strategy file on API
