+++ api/strategy.go
@@ -10,6 +10,7 @@ import (
 	"time"
 
 	"github.com/ethereum/go-ethereum/common"
+	queries "github.com/vocdoni/census3/db/sqlc"
 	"go.vocdoni.io/dvote/httprouter"
 	api "go.vocdoni.io/dvote/httprouter/apirest"
 	"go.vocdoni.io/dvote/log"
@@ -26,6 +27,31 @@ func (capi *census3API) initStrategiesHandlers() error {
 	}
 	return capi.endpoint.RegisterMethod("/strategy/token/{tokenID}", "GET",
 		api.MethodAccessTypePublic, capi.getTokenStrategies)
+}
+
+// createDummyStrategy creates the default strategy for a given token. This
+// basic strategy only includes the holders of the given token which have a
+// balance positive balance (holder_balance > 0).
+//
+// TODO: Only for the MVP, remove it.
+func (capi *census3API) createDummyStrategy(tokenID []byte) error {
+	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+	res, err := capi.sqlc.CreateStategy(ctx, "test")
+	if err != nil {
+		return err
+	}
+	strategyID, err := res.LastInsertId()
+	if err != nil {
+		return err
+	}
+	_, err = capi.sqlc.CreateStrategyToken(ctx, queries.CreateStrategyTokenParams{
+		StrategyID: strategyID,
+		TokenID:    tokenID,
+		MinBalance: big.NewInt(0).Bytes(),
+		MethodHash: []byte("test"),
+	})
+	return err
 }
 
 // getStrategies function handler returns the current registered strategies from
